- Change `productPrice` in `ShopProduct` from double to BigDecimal 	- Refactor all usages of `productPrice` to reflect data type change - Add `getListOfShops` method in `ShopService` that returns `List<Shop>` - Change `Shop` constructor to include parameter `products` of type `List<ShopProduct>` - Rename `Shop` methods `getMostExpensiveProduct` and `getCheapestProduct` to plural 	- Refactor all usages of methods to reflect name changes - Implement `Shop` methods `getMostExpensiveProducts` and `getCheapestProducts` that both return `List<ShopProduct>`

+++ src/main/java/ProductInterface.java
@@ -1,4 +1,6 @@
+import java.math.BigDecimal;
+
 public interface ProductInterface {
     String getProductName();
-    double getProductPrice();
+    BigDecimal getProductPrice();
 }
